+++ src/main/java/com/project/vetdata/controller/UserController.java
@@ -1,68 +1,18 @@
 package com.project.vetdata.controller;
 
-import com.project.vetdata.dto.UserCreateDTO;
-import com.project.vetdata.model.User;
-import com.project.vetdata.service.UserService;
-import io.swagger.v3.oas.annotations.Operation;
-import io.swagger.v3.oas.annotations.media.Content;
-import io.swagger.v3.oas.annotations.media.Schema;
-import io.swagger.v3.oas.annotations.responses.ApiResponse;
-import io.swagger.v3.oas.annotations.responses.ApiResponses;
-import jakarta.validation.Valid;
-import org.springframework.http.HttpStatus;
-import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.*;
+import org.springframework.stereotype.Controller;
+import org.springframework.ui.Model;
+import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.PathVariable;
+import org.springframework.web.bind.annotation.RequestMapping;
 
-import java.util.List;
-
-@RestController
-@RequestMapping("/users")
+@Controller
+@RequestMapping("users-pages")
 public class UserController {
-    private final UserService userService;
 
-    public UserController(UserService userService) {
-        this.userService = userService;
-    }
-
-    @Operation(summary = "Criar um novo usuário")
-    @ApiResponses(value = {
-            @ApiResponse(responseCode = "201", description = "Usuário criado com sucesso!",
-                    content = {@Content(mediaType = "application/json",
-                            schema = @Schema(implementation = User.class))}),
-            @ApiResponse(responseCode = "400", description = "Dados inválidos!", content = @Content)
-    })
-    @PostMapping
-    public ResponseEntity<User> createUser(@Valid @RequestBody UserCreateDTO userCreateDTO) {
-        User createdUser = userService.createUser(userCreateDTO);
-        return ResponseEntity.status(HttpStatus.CREATED).body(createdUser);
-    }
-
-    @Operation(summary = "Remover usuário por ID")
-    @ApiResponses(value = {
-            @ApiResponse(responseCode = "204", description = "Usuário removido!"),
-            @ApiResponse(responseCode = "404", description = "Usuário não encontrada!", content = @Content)
-    })
-    @DeleteMapping("/{id}")
-    public ResponseEntity<Void> deleteUser(@PathVariable Long id) {
-        return userService.getUserById(id)
-                .map(this::handleDelete)
-                .orElse(new ResponseEntity<>(HttpStatus.NOT_FOUND));
-    }
-
-    private ResponseEntity<Void> handleDelete(User user)  {
-        userService.deleteUser(user.getId());
-        return new ResponseEntity<>(HttpStatus.NO_CONTENT);
-    }
-
-    @Operation(summary = "Listar todos os usuários")
-    @ApiResponses(value = {
-            @ApiResponse(responseCode = "200", description = "Lista de usuários retornada com sucesso!",
-                    content = {@Content(mediaType = "application/json",
-                            schema = @Schema(implementation = User.class))})
-    })
-    @GetMapping
-    public ResponseEntity<List<User>> getAllUsers() {
-        List<User> users = userService.getAllUsers();
-        return ResponseEntity.ok(users);
+    @GetMapping("list")
+    public String list(Model model){
+        model.addAttribute("menuAtivo", "users");
+        return "user_list";
     }
 }
