+++ src/main/resources/templates/fragments/sidebar.html
@@ -25,10 +25,10 @@
         -->
 
         <ul>
-            <li class="active">
+            <li class="top-menu-invisible">
                 <a href="#" title="Dashboard"><i class="fa fa-lg fa-fw fa-bar-chart-o txt-color-blue"></i> <span class="menu-item-parent">Dashboard</span></a>
                 <ul>
-                    <li class="active">
+                    <li class="">
                         <a href="index.html" title="Dashboard"><span class="menu-item-parent">Analytics Dashboard</span></a>
                     </li>
                     <li class="">
@@ -36,16 +36,19 @@
                     </li>
                 </ul>
             </li>
-            <li class="top-menu-invisible">
+            <li th:class="${menuAtivo == 'breeds'} ? 'active open' : 'top-menu-invisible'">
                 <a href="#"><i class="fa fa-lg fa-fw  fa-paw txt-color-blue"></i> <span class="menu-item-parent">Raças</span></a>
                 <ul>
-                    <li class="">
+                    <li th:class="${submenuAtivo == 'list'} ? 'active'">
                         <a th:href="@{/breeds-pages/list}" title="Listar"><i class="fa fa-lg fa-fw  fa-list"></i> <span class="menu-item-parent">Listar</span></a>
                     </li>
-                    <li class="">
+                    <li th:class="${submenuAtivo == 'form'} ? 'active'">
                         <a th:href="@{/breeds-pages/form}" title="Nova Raça"><i class="fa fa-lg fa-fw  fa-save"></i> <span class="menu-item-parent">Cadastrar</span></a>
                     </li>
                 </ul>
+            </li>
+            <li th:class="${menuAtivo == 'users'} ? 'active'">
+                <a th:href="@{/users-pages/list}"><i class="fa fa-lg fa-fw  fa-user txt-color-blue"></i> <span class="menu-item-parent">Usuários</span></a>
             </li>
         </ul>
