+++ src/test/java/com/project/vetdata/controller/DiagnosticRestControllerIT.java
@@ -29,7 +29,7 @@ import static org.junit.jupiter.api.Assertions.*;
 
 @SpringBootTest(webEnvironment = SpringBootTest.WebEnvironment.RANDOM_PORT)
 @Testcontainers
-public class DiagnosticControllerIT {
+public class DiagnosticRestControllerIT {
 
     @LocalServerPort
     private Integer port;
