+++ src/test/java/com/project/vetdata/controller/UserRestControllerIT.java
@@ -31,7 +31,7 @@ import static org.junit.jupiter.api.Assertions.*;
 
 @SpringBootTest(webEnvironment = SpringBootTest.WebEnvironment.RANDOM_PORT)
 @Testcontainers
-public class UserControllerIT {
+public class UserRestControllerIT {
 
     @LocalServerPort
     private Integer port;
